+++ src/main/java/com/github/Naroru/JavaRushTelegramBot/command/basicCommands/NoCommand.java
@@ -1,5 +1,6 @@
-package com.github.Naroru.JavaRushTelegramBot.command;
+package com.github.Naroru.JavaRushTelegramBot.command.basicCommands;
 
+import com.github.Naroru.JavaRushTelegramBot.command.Command;
 import com.github.Naroru.JavaRushTelegramBot.service.SendMessageService;
 import org.telegram.telegrambots.meta.api.objects.Update;
 
@@ -7,7 +8,7 @@ public class NoCommand implements Command {
 
     private final SendMessageService sendMessageService;
 
-    private final String NO_MESSAGE = "Я поддерживаю команды, начинающиеся со слеша(/).\n"
+    public static final String NO_MESSAGE = "Я поддерживаю команды, начинающиеся со слеша(/).\n"
             + "Чтобы посмотреть список команд введите /help";
 
     public NoCommand(SendMessageService sendMessageService) {