+++ src/main/java/com/github/Naroru/JavaRushTelegramBot/command/basicCommands/StopCommand.java
@@ -6,7 +6,7 @@ import org.telegram.telegrambots.meta.api.objects.Update;
 
 public class StopCommand implements Command {
 
-    private final String STOP_MESSAGE = "Деактивировал все ваши подписки \uD83D\uDE1F.";
+    public static final  String STOP_MESSAGE = "Деактивировал все ваши подписки \uD83D\uDE1F.";
     private final SendMessageService sendMessageService;
 
     public StopCommand(SendMessageService sendMessageService) {