+++ src/test/java/com/github/Naroru/JavaRushTelegramBot/command/basicCommands/HelpCommandTest.java
@@ -1,0 +1,20 @@
+package com.github.Naroru.JavaRushTelegramBot.command.basicCommands;
+
+import com.github.Naroru.JavaRushTelegramBot.command.Command;
+import org.junit.jupiter.api.DisplayName;
+
+
+@DisplayName("Unit-testing HelpCommand")
+class HelpCommandTest extends AbstractCommandTest{
+
+
+    @Override
+     String getCommandMessage() {
+        return HelpCommand.HELP_MESSAGE;
+    }
+
+    @Override
+     Command getCommand() {
+        return new HelpCommand(sendMessageService);
+    }
+}