+++ src/test/java/com/github/Naroru/JavaRushTelegramBot/service/SendMessageServiceImpTest.java
@@ -1,6 +1,7 @@
 package com.github.Naroru.JavaRushTelegramBot.service;
 
 import com.github.Naroru.JavaRushTelegramBot.bot.JavarushBot;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -15,14 +16,40 @@ import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 @DisplayName("Unit-level testing for SendMessageService")
 @ExtendWith(MockitoExtension.class)
 class SendMessageServiceImpTest {
 
+    private SendMessageService sendMessageService;
 
+    @Mock
+    private JavarushBot javarushBot;
+
+    @BeforeEach
+    public void init()
+    {
+        sendMessageService = new SendMessageServiceImp(javarushBot);
+    }
+@Test
+    public void sendMessageShouldBeCorrect() throws TelegramApiException {
+    //given
+    String chatID = "123";
+    String message = "test message";
+
+    SendMessage sendMessage = new SendMessage();
+    sendMessage.setChatId(chatID);
+    sendMessage.setText(message);
+    sendMessage.enableHtml(true);
+
+    //when
+    sendMessageService.sendMessage(chatID, message);
+
+    //then
+    verify(javarushBot,only()).execute(sendMessage);
+
+}