+++ RELEASE_NOTES.md
@@ -3,3 +3,8 @@
 ## 0.1.0-SNAPSHOT
 * added stub telegram bot 
 * added Spring Boot skeleton
+* 
+## 0.2.0-SNAPSHOT
+* implemented Command pattern for the bot
+* added basic commands: /start /help /stop 
+* added actions for unknown command and for no command message